Move API views from api/views to sponsors/views
